remove user logged in event on AR use auth event stream + projection instead
