+++ config/packages/prooph_event_store.yaml
@@ -8,6 +8,9 @@ prooph_event_store:
             event_store: '@Prooph\EventStore\EventStore'
             connection:  '@app.pdo_connection'
             projections:
-                user_projection:
+                 auth_projection:
+                    read_model: App\Projection\Auth\AuthReadModel
+                    projection: App\Projection\Auth\AuthProjection
+                 user_projection:
                     read_model: App\Projection\User\UserReadModel
                     projection: App\Projection\User\UserProjection
