--- src/Model/User/Event/UserLoggedIn.php
@@ -1,23 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Model\User\Event;
-
-use App\Model\User\UserId;
-use Xm\SymfonyBundle\EventSourcing\AggregateChanged;
-
-final class UserLoggedIn extends AggregateChanged
-{
-    public static function now(UserId $userId): self
-    {
-        $event = self::occur($userId->toString());
-
-        return $event;
-    }
-
-    public function userId(): UserId
-    {
-        return UserId::fromString($this->aggregateId());
-    }
-}
