--- tests/Model/User/UserLoginTest.php
@@ -1,62 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\Model\User;
-
-use App\Model\User\Event;
-use App\Model\User\Exception;
-use App\Tests\BaseTestCase;
-
-class UserLoginTest extends BaseTestCase
-{
-    use UserTestTrait;
-
-    public function testLoggedIn(): void
-    {
-        $user = $this->getUserActive();
-
-        $user->loggedIn();
-
-        $events = $this->popRecordedEvent($user);
-
-        $this->assertRecordedEvent(
-            Event\UserLoggedIn::class,
-            [],
-            $events,
-        );
-
-        $this->assertCount(1, $events);
-    }
-
-    public function testLoggedInUnverified(): void
-    {
-        $user = $this->getUserActive(true);
-
-        $this->expectException(Exception\UserNotVerified::class);
-
-        $user->loggedIn();
-    }
-
-    public function testLoggedInInactive(): void
-    {
-        $user = $this->getUserInactive();
-
-        $this->expectException(Exception\InvalidUserActiveStatus::class);
-
-        $user->loggedIn();
-    }
-
-    public function testLoggedInDeleted(): void
-    {
-        $user = $this->getUserInactive();
-        $user->delete();
-
-        $this->expectException(Exception\UserIsDeleted::class);
-        $this->expectExceptionMessage(
-            sprintf('Tried to login deleted User with ID "%s"', $user->userId()),
-        );
-
-        $user->loggedIn();
-    }
-}
